translation_demo/requirements.txt
streamlit==1.35.0
whisper==1.1.10
torch==2.2.2
ffmpeg-python==0.2.0
deep-translator==1.11.4
gTTS==2.3.2
pyttsx3==2.90
pypiwin32==223
